chore: Refactor undo and redo functionality in Clicks/Undoredo.jsx and canvas.jsx
