+++ components/ButtonComponents/Clicks/Shapes.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Shapes = () => {
+    return (
+        <div>
+            <h1>Shapes</h1>
+        </div>
+    );
+}
+
+export default Shapes;
